// models/AgentAppStatus.ts
/**
 * Corellium API
 * REST API to manage your virtual devices.
 *
 * OpenAPI spec version: 6.0.0-20323
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* 
*/
export class AgentAppStatus {
    /**
    * 
    */
    'running'?: boolean | null;
    /**
    * 
    */
    'bundleID'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "running",
            "baseName": "running",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "bundleID",
            "baseName": "bundleID",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AgentAppStatus.attributeTypeMap;
    }

    public constructor() {
    }
}