// models/ApiNotFoundError.ts
/**
 * Corellium API
 * REST API to manage your virtual devices.
 *
 * OpenAPI spec version: 4.3.1-16664
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* 
*/
export class ApiNotFoundError {
    /**
    * Error text
    */
    'error': string;
    /**
    * Error ID
    */
    'errorID': string;
    /**
    * Name of the object requested
    */
    'name'?: string | null;
    /**
    * params provided by user
    */
    'params'?: any | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "error",
            "baseName": "error",
            "type": "string",
            "format": ""
        },
        {
            "name": "errorID",
            "baseName": "errorID",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "params",
            "baseName": "params",
            "type": "any",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiNotFoundError.attributeTypeMap;
    }

    public constructor() {
    }
}