// models/BtraceEnableOptions.ts
/**
 * Corellium API
 * REST API to manage your virtual devices.
 *
 * OpenAPI spec version: 5.7.1-19698
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* 
*/
export class BtraceEnableOptions {
    /**
    * 
    */
    'ranges'?: Array<Array<string>> | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "ranges",
            "baseName": "ranges",
            "type": "Array<Array<string>>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BtraceEnableOptions.attributeTypeMap;
    }

    public constructor() {
    }
}