// models/DomainOptions.ts
/**
 * Corellium API
 * REST API to manage your virtual devices.
 *
 * OpenAPI spec version: 5.5.0-18750
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TrialExtension } from '../models/TrialExtension';
import { HttpFile } from '../http/http';

/**
* 
*/
export class DomainOptions {
    /**
    * if true, totp is required
    */
    'totpRequired'?: boolean | null;
    'trialExtension'?: TrialExtension;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "totpRequired",
            "baseName": "totpRequired",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "trialExtension",
            "baseName": "trialExtension",
            "type": "TrialExtension",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DomainOptions.attributeTypeMap;
    }

    public constructor() {
    }
}